+++ tests/Integration/MSFT_XIISLogging.Integration.Tests.ps1
@@ -76,7 +76,7 @@ try
             { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should Not throw
         }
         #endregion
-        It 'Changing Loggging Truncate Settings ' -test {
+        It 'Changing Logging Truncate Settings ' -test {
 
             Invoke-Expression -Command "$($script:dscResourceName)_Truncate -OutputPath `$TestDrive"
             Start-DscConfiguration -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
@@ -85,6 +85,41 @@ try
 
             $currentLogSettings.directory | Should Be 'C:\IISLogFiles'
             $currentLogSettings.logExtFileFlags | Should Be 'Date,Time,ClientIP,UserName,ServerIP'
+            $currentLogSettings.logformat | Should Be 'W3C'
+            $currentLogSettings.logTargetW3C | Should Be 'File,ETW'
+            $currentLogSettings.TruncateSize | Should Be '2097152'
+            $currentLogSettings.localTimeRollover | Should Be 'True'
+            $currentLogSettings.customFields.Collection[0].LogFieldName | Should Be 'ClientEncoding'
+            $currentLogSettings.customFields.Collection[0].SourceName | Should Be 'Accept-Encoding'
+            $currentLogSettings.customFields.Collection[0].SourceType | Should Be 'RequestHeader'
+            $currentLogSettings.customFields.Collection[1].LogFieldName | Should Be 'X-Powered-By'
+            $currentLogSettings.customFields.Collection[1].SourceName | Should Be 'ASP.NET'
+            $currentLogSettings.customFields.Collection[1].SourceType | Should Be 'ResponseHeader'
+        }
+    }
+
+    Describe "$($script:dscResourceName)_LogFlags" {
+        #region DEFAULT TESTS
+        It 'Should compile without throwing' {
+            {
+                Invoke-Expression -Command "$($script:dscResourceName)_LogFlags -OutputPath `$TestDrive"
+                Start-DscConfiguration -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
+            } | Should not throw
+        }
+
+        It 'should be able to call Get-DscConfiguration without throwing' {
+            { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should Not throw
+        }
+        #endregion
+        It 'Changing enabled LogFlags ' -test {
+
+            Invoke-Expression -Command "$($script:dscResourceName)_LogFlags -OutputPath `$TestDrive"
+            Start-DscConfiguration -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
+
+            $currentLogSettings = Get-WebConfiguration -filter '/system.applicationHost/sites/siteDefaults/Logfile'
+
+            $currentLogSettings.directory | Should Be 'C:\IISLogFiles'
+            $currentLogSettings.logExtFileFlags | Should Be 'Date,Time,ClientIP,ServerIP,UserAgent'
             $currentLogSettings.logformat | Should Be 'W3C'
             $currentLogSettings.logTargetW3C | Should Be 'File,ETW'
             $currentLogSettings.TruncateSize | Should Be '2097152'